fixed conversion problems

+++ lib/include/lds/low_discr_seq_n.hpp
@@ -71,7 +71,7 @@ class sphere_n
 {
   private:
     vdcorput _vdc;
-    unsigned _n;
+    unsigned long _n;
     std::variant<std::unique_ptr<sphere_n>, std::unique_ptr<sphere>> _Sgen;
     double _range_t;
     double _t0;
